[#31] refactor: view 패키지가 dto 패키지와 독립적으로 구성되도록 수정
