+++ BE/baseball/src/main/java/kr/codesquad/baseball/web/view/InitialTeamView.java
@@ -1,4 +1,4 @@
-package kr.codesquad.baseball.web.dto.view;
+package kr.codesquad.baseball.web.view;
 
 import kr.codesquad.baseball.web.dto.PitcherDto;
 import kr.codesquad.baseball.web.dto.TeamDto;
